feat : authorization and people in group

+++ src/app/pages/discussion/one-discussion/one-discussion.component.ts
@@ -19,12 +19,12 @@ import {serverTimestamp} from '@angular/fire/firestore';
 })
 export class OneDiscussionComponent {
 
-  discussion: Discussion  = {
-    title:"",
-    creatorId:"",
-    participants:[],
-    createdAt:"",
-    id:""
+  discussion: Discussion = {
+    title: "",
+    creatorId: "",
+    participants: [],
+    createdAt: "",
+    id: ""
   };
   messages = <Message[]>[]
   newMessage: Message = {
@@ -35,7 +35,7 @@ export class OneDiscussionComponent {
     discussionId: "",
 
   }
-  currentUserId = signal<string | null>(null)
+  currentUserId: string = ""
 
   constructor(private discussionService: DiscussionService, private userService: UsersService, private router: Router, private messageService: MessageService) {
   }
@@ -49,7 +49,7 @@ export class OneDiscussionComponent {
       }
       console.log("discussion:", this.discussion)
       this.newMessage.discussionId = this.discussion.id
-      this.currentUserId.set(this.userService.getCurrentUser().uid)
+      this.currentUserId = this.userService.getCurrentUser().uid
       await this.getMessages()
 
 
@@ -62,7 +62,7 @@ export class OneDiscussionComponent {
       console.error("Erreur : discussionId non défini !");
       return;
     }
-    console.log("recuperation des messages pour :",this.discussion.id)
+    console.log("recuperation des messages pour :", this.discussion.id)
     let messages = await this.messageService.getAllMessageOfConversations(this.discussion.id)
     console.log("Messages récupérés :", messages);
     this.messages = messages
@@ -70,6 +70,14 @@ export class OneDiscussionComponent {
   }
 
   edit() {
+    if (this.currentUserId === "") {
+      console.log("no current user")
+      return;
+    }
+    if (this.currentUserId !== this.discussion.creatorId) {
+      console.log("unauthaurized")
+      return;
+    }
     this.discussionService.setRoute("edit-discussion")
   }
 
@@ -79,19 +87,71 @@ export class OneDiscussionComponent {
       console.log("no discussion")
       return;
     }
+
+    if (this.currentUserId === "") {
+      console.log("no current user")
+      return;
+    }
+    if (this.currentUserId !== this.discussion.creatorId) {
+      console.log("unauthaurized")
+      return;
+    }
+
     let res = await this.discussionService.deleteDiscussion(this.discussion)
 
     this.discussion = {
-      title:"",
-      creatorId:"",
-      participants:[],
-      createdAt:"",
-      id:""
+      title: "",
+      creatorId: "",
+      participants: [],
+      createdAt: "",
+      id: ""
     }
   }
 
-  async deleteMessage(messageId: string) {
-    let messageIdDeleted = await this.messageService.deleteMessage(messageId)
+  async leftDiscussion() {
+    if (this.currentUserId === "") {
+      console.log("no current user")
+      return;
+    }
+
+    if (this.currentUserId === this.discussion.creatorId) {
+      console.log("author cannot left own discussion")
+      return;
+    }
+    this.discussion.participants =  this.discussion.participants.filter(userId => userId !== this.currentUserId)
+    console.log("current user id :",this.currentUserId," and now the participants :",this.discussion.participants);
+    await this.discussionService.editDiscussion(this.discussion)
+
+    this.discussion = {
+      title: "",
+      creatorId: "",
+      participants: [],
+      createdAt: "",
+      id: ""
+    }
+     let discussions =  await this.discussionService.getDiscussions()
+  if (discussions && discussions.length > 0) {
+    this.discussionService.setCurrentDiscussion(discussions[0])
+
+  }
+      }
+
+  async deleteMessage(message: Message) {
+    if (this.currentUserId === "") {
+      console.log("no current user")
+      return;
+    }
+    if (!message.id) {
+      console.log("no message id")
+      return;
+    }
+
+    if (this.currentUserId !== message.creatorId) {
+      console.log("unauthaurized")
+      return;
+    }
+
+    let messageIdDeleted = await this.messageService.deleteMessage(message.id)
     this.messages = this.messages.filter(message => message.id !== messageIdDeleted);
   }
